Separate syntax test in single & multi variants
